requirements.txt
streamlit>=1.50.0
pandas
requests
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
